+++ app_plus/src/main/java/org/gkisalatiga/plus/lib/Logger.kt
@@ -53,6 +53,19 @@ class Logger {
         }
 
         /**
+         * Logging for whenever Firebase analytics logging is triggered.
+         * @param func Must be set to "{}" in order to correctly back-trace the caller method's enclosing class and method names.
+         * @param msg The message to be logged to the terminal.
+         * @param type The logging message type to be displayed (whether "verbose", "debug", "info", "error", or "warning").
+         */
+        fun logAnalytics (func: () -> Unit, msg: String, type: LoggerType = LoggerType.DEBUG) {
+            val tag = "$BASE_LOGGING_TAG-Analytics"
+            val msgString = "[${func.javaClass.enclosingClass?.name}.${func.javaClass.enclosingMethod?.name}] ::: $msg"
+
+            if (GlobalCompanion.DEBUG_ENABLE_LOG_CAT_ANALYTICS) doLog(tag, msgString, type)
+        }
+
+        /**
          * Logging for "on boot signal received"-related operations.
          * @param func Must be set to "{}" in order to correctly back-trace the caller method's enclosing class and method names.
          * @param msg The message to be logged to the terminal.